fix: update test mock paths for Phase 5 refactored modules
Update vi.mock paths in test files to point to the actual source modules
in features/ directories instead of the re-export stubs in core/.

Files fixed:
- labs.test.ts: mock features/labs/definitions/features
- useFeatureFlag.test.ts: mock features/labs/definitions/features
- LayoutListItem.test.tsx: mock features/layout-library LayoutActions
- LayoutList.test.tsx: mock features/layout-library LayoutActions
- MobileLayoutsPanel.test.tsx: add missing storage mock functions

All 3356 tests now pass.
