+++ src/test/hooks/useFeatureFlag.test.ts
@@ -3,10 +3,11 @@ import { renderHook } from '@testing-library/react';
 import { useFeatureFlag, isFeatureEnabled } from '../../hooks/useFeatureFlag';
 import { useLabsStore } from '../../core/store/labs';
 import { resetAllStores } from '../testUtils';
-import * as features from '../../core/labs/features';
+import * as features from '../../features/labs/definitions/features';
 
-vi.mock('../../core/labs/features', async () => {
-  const actual = await vi.importActual<typeof features>('../../core/labs/features');
+// Must mock the actual source module that the store imports from
+vi.mock('../../features/labs/definitions/features', async () => {
+  const actual = await vi.importActual<typeof features>('../../features/labs/definitions/features');
   return {
     ...actual,
     getFeature: vi.fn(),
